# auth_service/docs/swagger.yaml
definitions:
  handler.AuthHandler:
    type: object
info:
  contact: {}
paths:
  /login:
    post:
      consumes:
      - application/json
      description: |-
        Вход пользователя в систему и получение JWT-токена
        Поля:
        - username: Логин (обязательно)
        - password: Пароль (мин. 8 символов)
        - email: Email (должен быть валидным)
      parameters:
      - description: Данные для входа
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.AuthHandler'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              id:
                type: integer
              message:
                type: string
              token:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Аутентификация пользователя
      tags:
      - auth
  /register:
    post:
      consumes:
      - application/json
      description: |-
        Регистрирует нового пользователя в системе
        Поля:
        - username: Логин (обязательно)
        - password: Пароль (мин. 8 символов)
        - email: Email (должен быть валидным)
        - repeat_password: Должен совпадать с password
      parameters:
      - description: Данные для регистрации
        in: body
        name: input
        required: true
        schema:
          $ref: '#/definitions/handler.AuthHandler'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            properties:
              id:
                type: integer
              message:
                type: string
            type: object
        "400":
          description: Bad Request
          schema:
            properties:
              error:
                type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            properties:
              error:
                type: string
            type: object
      summary: Регистрация пользователя
      tags:
      - auth
swagger: "2.0"